requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
django-storages[azure]
azure-storage-blob
Pillow
PyPDF2
requests
reportlab